+++ ft-plugin/src/main/java/com/ft/plugin/FTPlugin.java
@@ -65,7 +65,8 @@ public class FTPlugin implements Plugin<Project> {
             PluginConfigManager.get().setExtension(extension);
 
             Logger.setDebug(extension.showLog);
-            Logger.debug("ASM Version: asm7");
+            Logger.debug("Plugin Version:" + BuildConfig.PLUGIN_VERSION +
+                    ",ASM Version:asm7");
 
             FTMapUploader f = new FTMapUploader(p, extension);
             f.configMapUpload();